+++ api/modules/payoutpt/module.go
@@ -4,6 +4,7 @@ import (
 	"hrms/modules/payoutpt/internal/feature/cancel"
 	"hrms/modules/payoutpt/internal/feature/create"
 	"hrms/modules/payoutpt/internal/feature/get"
+	"hrms/modules/payoutpt/internal/feature/haspending"
 	"hrms/modules/payoutpt/internal/feature/list"
 	"hrms/modules/payoutpt/internal/feature/pay"
 	"hrms/modules/payoutpt/internal/repository"
@@ -12,23 +13,24 @@ import (
 	"hrms/shared/common/mediator"
 	"hrms/shared/common/middleware"
 	"hrms/shared/common/module"
+	"hrms/shared/contracts"
 
 	"github.com/gofiber/fiber/v3"
 )
 
 type Module struct {
-	ctx        *module.ModuleContext
-	repo       repository.Repository
-	tokenSvc   *jwt.TokenService
-	eb         eventbus.EventBus
+	ctx      *module.ModuleContext
+	repo     repository.Repository
+	tokenSvc *jwt.TokenService
+	eb       eventbus.EventBus
 }
 
 func NewModule(ctx *module.ModuleContext, tokenSvc *jwt.TokenService) *Module {
 	repo := repository.NewRepository(ctx.DBCtx)
 	return &Module{
-		ctx:        ctx,
-		repo:       repo,
-		tokenSvc:   tokenSvc,
+		ctx:      ctx,
+		repo:     repo,
+		tokenSvc: tokenSvc,
 	}
 }
 
@@ -41,6 +43,10 @@ func (m *Module) Init(eb eventbus.EventBus) error {
 	mediator.Register[*get.Query, *get.Response](get.NewHandler())
 	mediator.Register[*pay.Command, *pay.Response](pay.NewHandler())
 	mediator.Register[*cancel.Command, mediator.NoResponse](cancel.NewHandler())
+
+	// Contract handlers for cross-module communication
+	mediator.Register[*contracts.HasPendingPayoutPTQuery, *contracts.HasPendingPayoutPTResponse](haspending.NewHandler(m.repo))
+
 	return nil
 }